Adding likes and dislikes

+++ app/Http/Controllers/LikeController.php
@@ -15,7 +15,6 @@ class LikeController extends Controller
             return redirect('/login');
         }
 
-      //  $post = Post::where('id', $postId)->get()->first();
         $curUser = User::where('id', session()->get('userId'))->get()->first();
 
         $record = Like::where('userId', $curUser->id)->where('postId', $postId)->first();
@@ -26,6 +25,27 @@ class LikeController extends Controller
             $newLike->postId = $postId;
             $newLike->like = true;
             $newLike->dislike = false;
+            $newLike->save();
+        } else {
+            $record->delete();
+        }
+
+        return Redirect::back();
+    }
+
+    public function dislikePost($postId) {
+        if(!session()->get('userId')) {
+            return redirect('/login');
+        }
+        $curUser = User::where('id', session()->get('userId'))->get()->first();
+        $record = Like::where('userId', $curUser->id)->where('postId', $postId)->first();
+
+        if(!isset($record)) {
+            $newLike = new Like;
+            $newLike->userId = $curUser->id;
+            $newLike->postId = $postId;
+            $newLike->like = false;
+            $newLike->dislike = true;
             $newLike->save();
         } else {
             $record->delete();
